BackEnd/AuthApi-drf/auth_api01/requirements.txt
asgiref==3.8.1
Django==5.1.2
sqlparse==0.5.1
tzdata==2024.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pillow==10.4.0
PyJWT==2.9.0